[18] Defined simple model design. Added User, Organization, Volunteer, Project, ProjectVolunteer (Relation object) fields. Added UserType.cs enumerator. Filled Database context file ready for migrations

+++ server/WebAPI/Base/Interfaces/IBaseEntity.cs
@@ -2,6 +2,6 @@
 {
     public interface IBaseEntity
     {
-        int Id { get; set; }
+        long Id { get; set; }
     }
 }
